fix/calcul_arrondi (#90)

+++ src/main/java/fr/formationacademy/scpiinvestplusapi/dto/InvestmentStatisticsDtoOut.java
@@ -8,9 +8,9 @@ import java.util.Map;
 @Data
 @Builder
 public class InvestmentStatisticsDtoOut {
-    public int montantInvesti;
+    public double montantInvesti;
     public double rendementMoyen;
-    public int revenuMensuel;
+    public double revenuMensuel;
     public double cashbackMontant;
     private Map<String, Double> repGeographique;
     private Map<String, Double> repSectoriel;
